Port mouse.rb to C
This is more or less a direct port.

I addressed default arguments by creating multiple
versions of the same method with a number appended
to the end. A higher number always means more args
are expected. I tried to be consistent about what
arguments should be given and the order, but it
still needs a bit of work.

Documentation is also a WIP.